requirements.txt
requests
beautifulsoup4
pandas
lxml
python-dotenv
matplotlib